Formula 'templates' and disabled 'Add' instruction
Instead of creating formulas again and again,
translator now uses "formula 'templates'".
Those are created once and when needed to use,
they are only parametrized by variables or constants.
For this reason some classes was renamed to 'Abstract*'
and 'Concrete*'. Naming is not uniformed, so it will probably
change in future.

Because 'Add' instruction is not as simple as others,
it was (temporarily) disabled.